fix: deactivate profile compression and fix bug

+++ handlers/baseHandler.ts
@@ -3,7 +3,7 @@ import { PrismaClient } from "@prisma/client";
 import { PrismaInstance } from "@services"
 import { StorageInstance } from "services/storage";
 import { randomUUID } from "crypto";
-import sharp from "sharp";
+// import sharp from "sharp";
 
 export abstract class BaseHandler {
     protected prisma: PrismaClient;
@@ -21,7 +21,8 @@ export abstract class BaseHandler {
             const imageBuffer = Buffer.from(base64Image, "base64");
 
             // compress image buffer
-            const compressedImageBuffer = await sharp(imageBuffer).resize(150, 150).toBuffer();
+            // const compressedImageBuffer = await sharp(imageBuffer).resize(150, 150).toBuffer();
+            const compressedImageBuffer = imageBuffer
 
             // Get file extension {jpeg, png}
             const terminator = pictureData.indexOf(";");
@@ -57,6 +58,7 @@ export abstract class BaseHandler {
 
     protected deletePictureFile = async (filePath: string) => {
         try {
+            if(filePath === "") return;
             const file = this.bucket.file(filePath);
             await file.delete();
